requirements.txt
kafka-python
psycopg2-binary
python-dotenv
faker
pandas